requirements.txt
alembic==1.14.1
babel==2.17.0
blinker==1.8.2
click==7.1.2
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==20.1.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.3.10
MarkupSafe==2.0.1
passlib==1.7.4
python-dotenv==1.0.1
pytz==2025.2
speaklater==1.3
SQLAlchemy==1.3.24
typing_extensions==4.13.2
Werkzeug==1.0.1
WTForms==3.1.2
zipp==3.20.2
